Move some panels to extras
